October 2023 Update
Uploading changes from local repo.

+++ Python/Kindle Scribe/scribe notebook pdf converter.py
@@ -1,7 +1,8 @@
-# This script requires the use of a the KFX-Input plugin for Calibre, found on Mobileread.
-# Convert Kindle Scribe notebook files into epub, and later pdf files. 
+# Requirements: This script requires the use of a the KFX-Input plugin for Calibre, found on Mobileread.
+#
+# Usage: Convert Kindle Scribe notebook files into epub, and later pdf files. 
 # Margins are set to a half inch, passed though in the calibre command.
-# August 10, 2023
+# October 1, 2023
 
 import os
 import subprocess
@@ -11,24 +12,32 @@ calibre_debug_executable = r"C:\Program Files\Calibre2\calibre-debug.exe"
 calibre_convert_executable = r"C:\Program Files\Calibre2\ebook-convert.exe"
 
 # Path to the directories. Where notebook_directory is where they are stored on Kindle.
-notebook_directory = r"D:\<Scribe root folder>\.notebooks"
-output_directory = r"C:\<Scribe root folder>\Kindle_Scribe_PDF_Files"
+notebook_directory = r"C:\<Scribe root folder>\.notebooks"
+output_directory = r"C:\<PC root folder>\Kindle_Scribe_PDF_Files"
+
+# Temporary directory for epub files
+epub_directory = output_directory + r"\scribePython_temp_epub_files"
+if not os.path.exists(epub_directory):
+    os.mkdir(epub_directory)
+
 
 # Loop through each folder in the base directory
+print("Script will begin processing your files...")
+
 for folder_name in os.listdir(notebook_directory):
     folder_path = os.path.join(notebook_directory, folder_name)
     
     # Check if the item is a directory. Don't run if not a notebook directory
     # Thumbnail, and personal document files are excluded.
-    if os.path.isdir(folder_path) and folder_name != ".backup" and folder_name != "thumbnails" and "!!PDOC!!" not in folder_name and "!!EBOK!!" not in folder_name:
+    if os.path.isdir(folder_path) and len(folder_name) == 36 and folder_name != ".backup" and folder_name != "thumbnails" and "!!PDOC!!" not in folder_name and "!!EBOK!!" not in folder_name:
         # Construct the inital command to convert to ePub.
-        command = [calibre_debug_executable, "-r", "KFX Input", "--", folder_path, f"{folder_name}.epub"]
+        command = [calibre_debug_executable, "-r", "KFX Input", "--", folder_path, os.path.join(epub_directory, f"{folder_name}.epub")]
         
         # Run the command
         subprocess.run(command)
         
         # Construct the second command to convert ePub to PDF.
-        epub_file = os.path.join(notebook_directory, f"{folder_name}.epub")
+        epub_file = os.path.join(epub_directory, f"{folder_name}.epub")
         pdf_file = os.path.join(output_directory, f"{folder_name}.pdf")
         ebook_convert_command = [
             calibre_convert_executable,
@@ -44,16 +53,17 @@ for folder_name in os.listdir(notebook_directory):
         subprocess.run(ebook_convert_command)
         
 # Prompt the user if they would like to delete the intermediary ePub files.
-user_input = input("Would you like to delete .epub files? (Y/N): ").strip().lower()
+user_input = input("Would you like to delete .epub files? WARNING THIS WILL DELETE ALL EPUB FILES IN 'scribe_temp_epub_files' DIRECTORY (Y/N): ").strip().lower()
 if user_input == "y":
     try:
         # Loop through all ePub files in notebook directory, and delete.
-        for epub_file in os.listdir(notebook_directory):
-        	if epub_file.endswith(".epub"):
-                    file_path = os.path.join(notebook_directory, epub_file)
+        for epub_file in os.listdir(epub_directory):
+			# As of 2023, the filename for scribe notebook files are 41 characters long.
+        	if epub_file.endswith(".epub") and len(epub_file) == 41:
+                    file_path = os.path.join(epub_directory, epub_file)
                     os.remove(file_path)
                     print(f"Deleted: {file_path}")
-                print("Deletion process completed.")
+                    print("Deletion process completed. Script has completed.")
     except Exception as e:
         print(f"An error occurred: {e}")
 else:
